+++ libMx/mx_timer.c
@@ -7,7 +7,7 @@
  *
  *--------------------------------------------------------------------------
  *
- * Copyright 1999, 2001-2002, 2004 Illinois Institute of Technology
+ * Copyright 1999, 2001-2002, 2004, 2006 Illinois Institute of Technology
  *
  * See the file "LICENSE" for information on usage and redistribution
  * of this file, and for a DISCLAIMER OF ALL WARRANTIES.
@@ -282,7 +282,7 @@ mx_timer_read( MX_RECORD *timer_record, double *seconds )
 }
 
 MX_EXPORT mx_status_type
-mx_timer_get_mode( MX_RECORD *timer_record, int *mode )
+mx_timer_get_mode( MX_RECORD *timer_record, long *mode )
 {
 	static const char fname[] = "mx_timer_get_mode()";
 
@@ -309,7 +309,7 @@ mx_timer_get_mode( MX_RECORD *timer_record, int *mode )
 }
 
 MX_EXPORT mx_status_type
-mx_timer_set_mode( MX_RECORD *timer_record, int mode )
+mx_timer_set_mode( MX_RECORD *timer_record, long mode )
 {
 	static const char fname[] = "mx_timer_set_mode()";
